+++ src/test/java/domain/piece/GuardTest.java
@@ -1,9 +1,7 @@
-package domain;
+package domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import domain.piece.Guard;
-import domain.piece.PieceFactory;
 import domain.position.Distance;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
